Update msbuild to include additional files (.ini files) and automatically copy dinput8.dll with the correct name
This update to the msbuild workflow should facilitate the fix installation for the end-user. New projects must create a dummy dll file with the naming convention "$(ProjectName)_dinput8.dll" in the release output directory, which will then be used by the script to copy the asi loader with the correct name.

For example, if a certain game needs winmm.dll, the complete command that should be used for the post build is:
"copy  NUL $(OutDirFullPath)$(ProjectName)_winmm.dll"

The same naming convention applies to include additional files (e.g. .ini)